+++ src/Handlers/GameMessage.cs
@@ -36,9 +36,6 @@ namespace Amnesia.Handlers
 
                 if (Config.ProtectMemoryDuringPvp && mainName != secondaryName)
                 {
-                    // TODO: this is nice, but damage/kill handling needs to also be redone to include the killing player in game message even if that player is offline
-                    //  and probably also to give that player offline credit for the kill(s).
-
                     var killerClient = ConnectionManager.Instance.Clients.GetForPlayerName(secondaryName);
                     if (killerClient != null)
                     {